requirements.txt
fastapi
uvicorn
streamlit>=1.45.0
python-dotenv
pydantic
requests
beautifulsoup4
pandas
numpy
plotly
httpx
neo4j
openai
scikit-learn
warcio
tqdm
langdetect
apscheduler